+++ src/types/PortmanConfig.ts
@@ -90,6 +90,7 @@ export type ResponseBodyTest = {
   length?: string | number
   minLength?: string | number
   maxLength?: string | number
+  notExist?: boolean
 }
 
 export type ResponseHeaderTest = {
@@ -100,6 +101,7 @@ export type ResponseHeaderTest = {
   length?: string | number
   minLength?: string | number
   maxLength?: string | number
+  notExist?: boolean
 }
 
 type OverwriteConfig = {