+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
-- Overwrites - Automatically convert values for request query parameters, path variables & headers to string ((#336).
+- Overwrites - Automatically convert values for request query parameters, path variables & headers to string (#336)
+- Fix ContentCheck with 0 (zero) for length/minLength/maxLength (#369)
 
 ## v1.17.0 - (2022-06-21)